requirements.txt
PyQt6>=6.4.0
numpy>=1.24.0
# BiRefNet: use Hugging Face weights + transformers (see core/birefnet.py). Optional: accelerate, timm — add only if model load fails.
transformers>=4.38.0
huggingface_hub>=0.24.0
einops>=0.7.0
kornia>=0.7.0
timm>=0.9.0
socksio>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
pytoshop>=1.2.0
pytest>=7.0.0
